requirements.txt
Flask==2.0.3
transformers==4.6.1
sentence-transformers==2.1.0
torch==1.8.1
openai==0.8.2
numpy==1.20.3
faiss-cpu==1.7.0
